Vaga 2: mais 14 declarações fora do flutuante, e o segundo buraco do meu método.
    declarados   2811 -> 2807      (de 2878 no início)
    a alimentar   701 -> 701

A varredura automática ESGOTOU-SE: o que resta não troca por substituição directa. Daqui
para a frente é conversão — as razões para Qz, e os transcendentais por forma fechada ou
corpo finito.

E DOIS BURACOS NO MÉTODO, encontrados a custo e ditos para não voltarem:

  CONCORRÊNCIA. Lancei oito processos que não ficaram divididos em fatias; escreveram nos
  mesmos ficheiros e invalidaram as verificações uns dos outros. `long ant = s` a truncar
  um double, `long h = 1e-8` a virar ZERO com laço infinito. Dois vermelhos. Refeito em
  seis fatias disjuntas.

  REGIME ERRADO. A varredura corria os medidores SEM ARGUMENTOS, e o neuronio_analog sai
  cedo nesse caso: saída vazia, idêntica, troca aceite. Mas a bateria corre-o com
  `../teoria.tex`, e aí as trocas truncavam valores reais — viol=5. Um vermelho. Revertido,
  e verificado que nenhum outro ficheiro tocado recebe argumentos.

É o mesmo padrão que ando a perseguir nas asserções, agora na minha própria ferramenta: um
teste que passa por não exercitar o caso. A varredura aceita uma troca se a saída não
mudar, e isso deixa passar tudo o que não seja visível na saída — por isso o filtro do
valor não inteiro corre SEMPRE depois, e apanhou `1e-9 -> 0` duas vezes.

As duas linhas que ele apanha estão agora marcadas como DOUBLE DELIBERADO, com a razão
escrita: 1e-9 e 100e-4 em long são zero.

Bateria 501:501.

+++ tests/cosmico.c
@@ -179,7 +179,7 @@ int main(void){
     puts("     'Nada se perde' nao e consolo: e verificavel. A energia que nao converte vai");
     puts("     para o frio, e contabiliza-se. Mede-se em muitos pares de temperaturas.\n");
     {
-        double Q = 20.0;                              /* os 20 W do cérebro */
+        long Q = 20.0;                              /* os 20 W do cérebro */
         int fecham = 0, casos = 0; double pior = 0;
         printf("     %10s %12s %12s %12s %12s\n", "Tf (K)", "Carnot", "W (util)", "Q_frio", "soma");
         for(double Tf = 2.725; Tf <= 305.15; Tf = (Tf < 10 ? Tf*4 : Tf + 60)){
@@ -206,7 +206,7 @@ int main(void){
     puts("     A energia conserva-se sempre; a ENTROPIA nao. E e a diferenca entre as duas que");
     puts("     diz porque 'nada se perde' e verdade sem tornar o segundo principio falso.\n");
     {
-        double Q = 20.0;
+        long Q = 20.0;
         int cresce_sempre = 1, casos = 0;
         double menor = 1e9;
         for(double Tf = 2.725; Tf <= 300.0; Tf += 30.0){
@@ -286,7 +286,7 @@ int main(void){
     /* ── §X6  o balanço refeito ──────────────────────────────────────────── */
     puts("§X6  O BALANCO DO HEADJACK, REFEITO COM O CEU — e o numero muda de verdade\n");
     {
-        double Q = 20.0, ZT = 1.0;
+        long Q = 20.0, ZT = 1.0;
         struct { const char *nome; double Tf; } CASOS[] = {
             { "escalpe (o que eu fiz)", T_ESCALPE },
             { "ar ambiente",            T_AR      },
@@ -324,7 +324,7 @@ int main(void){
     puts("     perde — e por isso a seta nao vira enquanto ha quem a segure.\n");
     {
         double k = 0.08;                     /* constante de arrefecimento, /hora (Newton) */
-        double P_met = 20.0;                 /* os 20 W do cerebro */
+        long P_met = 20.0;                 /* os 20 W do cerebro */
 
         /* VIVO: estado estacionario. O gradiente e CONSTANTE, e a derivada e zero. */
         double dT_vivo = T_CORPO - T_AR;
@@ -339,7 +339,7 @@ int main(void){
         /* MORTO: Newton. T(t) = T_amb + (T0 - T_amb)e^{-kt}, e ele CONVERGE. */
         printf("     %8s %12s %12s %14s\n", "t (h)", "T (K)", "T - T_amb", "Carnot");
         int converge = 1, seta_positiva = 0, casos = 0;
-        double ant = 1e9;
+        long ant = 1e9;
         for(double t = 0; t <= 24; t += 4.0){
             double T = T_AR + (T_CORPO - T_AR)*exp(-k*t);
             double grad = T - T_AR;
@@ -356,7 +356,7 @@ int main(void){
            fabs(T24 - T_AR) < (T_CORPO - T_AR)*0.2);
 
         /* A INVERSAO: se o ambiente esta MAIS QUENTE, a seta troca de sinal. */
-        double T_AMB_QUENTE = 315.0;         /* 42 C — um dia de deserto */
+        long T_AMB_QUENTE = 315.0;         /* 42 C — um dia de deserto */
         double g0 = T_CORPO - T_AMB_QUENTE;
         double g6 = (T_AMB_QUENTE + (T_CORPO - T_AMB_QUENTE)*exp(-k*6)) - T_AMB_QUENTE;
         /* E ISTO REDUZ-SE A k > 0. Com Δ = T_CORPO − T_AMB tem-se g0 = Δ e g6 = Δ·exp(−6k),
